Rename monitoring to metrics_collector
Longhorn 685
