+++ metrics_collector/base_collector.go
@@ -1,4 +1,4 @@
-package monitoring
+package metricscollector
 
 import (
 	"github.com/sirupsen/logrus"
